move receiptToResultLine to adventureOrcustrator
